[HttpKernel] removed deprecated profiler storages
