+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/profiling.xml
@@ -11,11 +11,8 @@
             <argument type="service" id="logger" on-invalid="null" />
         </service>
 
-        <service id="profiler.storage" class="" public="false">
+        <service id="profiler.storage" class="Symfony\Component\HttpKernel\Profiler\FileProfilerStorage" public="false">
             <argument>%profiler.storage.dsn%</argument>
-            <argument>%profiler.storage.username%</argument>
-            <argument>%profiler.storage.password%</argument>
-            <argument>%profiler.storage.lifetime%</argument>
         </service>
 
         <service id="profiler_listener" class="Symfony\Component\HttpKernel\EventListener\ProfilerListener">
